pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cognee-framework"
version = "1.0.0"
description = "AI-Powered Prompt Development and Evaluation System using Cognee Knowledge Graphs"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Claude Code Prompt Development Framework"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.8"
dependencies = [
    "claude-code-sdk>=0.0.10",
    "cognee>=0.1.43",
    "anthropic>=0.25.0",
    "requests>=2.31.0",
    "sentence-transformers>=2.2.2",
    "rouge-score>=0.1.2",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "pandas>=2.0.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "matplotlib>=3.7.0",
    "plotly>=5.15.0",
    "seaborn>=0.12.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.65.0",
    "jinja2>=3.1.0",
    "scikit-learn>=1.3.0",
    "statsmodels>=0.14.0",
    "nltk>=3.8.0",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0.0",
    "ipykernel>=6.23.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]
nlp = [
    "spacy>=3.6.0",
]

[project.urls]
Homepage = "https://github.com/your-username/cognee-framework"
Repository = "https://github.com/your-username/cognee-framework"
Documentation = "https://github.com/your-username/cognee-framework/docs"

[project.scripts]
cognee-evaluate = "scripts.evaluate:main"
cognee-compare = "scripts.compare:main"

[tool.setuptools.packages.find]
include = ["cognee_framework*"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503"]